(ISSUE-23) Model Tink. In this commit, two primitives have been modeled:
   * AEAD Ciphers
   * MAC

Note that this specification also models the implementation of the
Google Tink API for generating KeyTemplates (see the modules AeadKeyTemplates
and MacKeyTemplates). Using this approach, it was not necessary to
change the CrySL language (considering some ideas that arose in a discussion with Johannes,
which I tried to follow). Nevertheless, I am still not sure if we should
specify the implementation of the library, instead of only modeling
how to use an API. In addition, I only specified the standard keys
supported by the library. Developers could use some factory methods
to generate keys, but I am not sure if this approach is secure.