As per @rakshitkr 's suggestion: Fixing cipher rule to avoid FNs.
In a first step, CogniCrypt_SAST's constraint solver filters out
non-applicable constraints. It thereby considers any constraint that
refers to an object CogniCrypt_SAST has not previously encountered in
the target program as non-applicable. With the two instanceof()
constraints, this lead to a situations where (incomplete) programs
without init calls would not result in an error message even when an
insecure algorithm was used. Adding this one line fixes that.